sql: Improve testing and documentation for constantFolderVisitor
This change pulls the constant related visitors into constant.go and
cleans up a few tests.
