+++ sql/parser/normalize.go
@@ -18,8 +18,6 @@ package parser
 
 import (
 	"fmt"
-	"go/constant"
-	"go/token"
 	"strings"
 )
 
@@ -527,134 +525,4 @@ func ContainsVars(expr Expr) bool {
 	v := containsVarsVisitor{containsVars: false}
 	WalkExprConst(&v, expr)
 	return v.containsVars
-}
-
-type constantFolderVisitor struct{}
-
-var _ Visitor = &constantFolderVisitor{}
-
-func (*constantFolderVisitor) VisitPre(expr Expr) (recurse bool, newExpr Expr) {
-	return expr != nil, expr
-}
-
-var unaryOpToToken = map[UnaryOperator]token.Token{
-	UnaryPlus:  token.ADD,
-	UnaryMinus: token.SUB,
-}
-var unaryOpToTokenIntOnly = map[UnaryOperator]token.Token{
-	UnaryComplement: token.XOR,
-}
-var binaryOpToToken = map[BinaryOperator]token.Token{
-	Plus:  token.ADD,
-	Minus: token.SUB,
-	Mult:  token.MUL,
-	Div:   token.QUO, // token.QUO_ASSIGN to force integer division.
-}
-var binaryOpToTokenIntOnly = map[BinaryOperator]token.Token{
-	Mod:    token.REM,
-	Bitand: token.AND,
-	Bitor:  token.OR,
-	Bitxor: token.XOR,
-}
-var binaryShiftOpToToken = map[BinaryOperator]token.Token{
-	LShift: token.SHL,
-	RShift: token.SHR,
-}
-var comparisonOpToToken = map[ComparisonOp]token.Token{
-	EQ: token.EQL,
-	NE: token.NEQ,
-	LT: token.LSS,
-	LE: token.LEQ,
-	GT: token.GTR,
-	GE: token.GEQ,
-}
-
-func (*constantFolderVisitor) VisitPost(expr Expr) (retExpr Expr) {
-	defer func() {
-		// constant operations can panic for a number of reasons. It's
-		// safest to just recover here.
-		if r := recover(); r != nil {
-			retExpr = expr
-		}
-	}()
-	switch t := expr.(type) {
-	case *ParenExpr:
-		if cv, ok := t.Expr.(*NumVal); ok {
-			return cv
-		}
-	case *UnaryExpr:
-		if cv, ok := t.Expr.(*NumVal); ok {
-			if token, ok := unaryOpToToken[t.Operator]; ok {
-				return &NumVal{Value: constant.UnaryOp(token, cv.Value, 0)}
-			}
-			if cv.Kind() == constant.Int {
-				if token, ok := unaryOpToTokenIntOnly[t.Operator]; ok {
-					return &NumVal{Value: constant.UnaryOp(token, cv.Value, 0)}
-				}
-			}
-		}
-	case *BinaryExpr:
-		l, okL := t.Left.(*NumVal)
-		r, okR := t.Right.(*NumVal)
-		if okL && okR {
-			if token, ok := binaryOpToToken[t.Operator]; ok {
-				return &NumVal{Value: constant.BinaryOp(l.Value, token, r.Value)}
-			}
-			if l.Kind() == constant.Int && r.Kind() == constant.Int {
-				if token, ok := binaryOpToTokenIntOnly[t.Operator]; ok {
-					return &NumVal{Value: constant.BinaryOp(l.Value, token, r.Value)}
-				}
-				if rInt, err := r.asInt(); err == nil && rInt >= 0 {
-					if token, ok := binaryShiftOpToToken[t.Operator]; ok {
-						return &NumVal{Value: constant.Shift(l.Value, token, uint(rInt))}
-					}
-				}
-			}
-		}
-	case *ComparisonExpr:
-		l, okL := t.Left.(*NumVal)
-		r, okR := t.Right.(*NumVal)
-		if okL && okR {
-			if token, ok := comparisonOpToToken[t.Operator]; ok {
-				return MakeDBool(DBool(constant.Compare(l.Value, token, r.Value)))
-			}
-		}
-	}
-	return expr
-}
-
-func foldNumericConstants(expr Expr) (Expr, error) {
-	// TODO(nvanbenschoten) Investigate normalizing associative operations to group
-	// constants together and permit further numeric constant folding.
-	v := constantFolderVisitor{}
-	expr, _ = WalkExpr(&v, expr)
-	return expr, nil
-}
-
-type constantTypeVisitor struct {
-	cfv constantFolderVisitor
-}
-
-var _ Visitor = &constantTypeVisitor{}
-
-func (*constantTypeVisitor) VisitPre(expr Expr) (recurse bool, newExpr Expr) {
-	switch t := expr.(type) {
-	case Constant:
-		typedConst, err := t.TypeCheck(nil, nil)
-		if err != nil {
-			panic(err)
-		}
-		return false, typedConst
-	}
-	return expr != nil, expr
-}
-
-func (*constantTypeVisitor) VisitPost(expr Expr) (retExpr Expr) { return expr }
-
-// TypeConstants TODO(nvanbenschoten) MORTY
-func TypeConstants(expr Expr) (Expr, error) {
-	v := constantTypeVisitor{}
-	expr, _ = WalkExpr(&v.cfv, expr)
-	expr, _ = WalkExpr(&v, expr)
-	return expr, nil
 }
