+++ sql/parser/scan_test.go
@@ -256,7 +256,7 @@ func TestScanNumber(t *testing.T) {
 func TestScanParam(t *testing.T) {
 	testData := []struct {
 		sql      string
-		expected int
+		expected int64
 	}{
 		{`$1`, 1},
 		{`$1a`, 1},
@@ -269,7 +269,7 @@ func TestScanParam(t *testing.T) {
 		if id != PARAM {
 			t.Errorf("%s: expected %d, but found %d", d.sql, PARAM, id)
 		}
-		i, err := lval.union.numVal().asInt()
+		i, err := lval.union.numVal().asInt64()
 		if err != nil {
 			t.Errorf("%s: expected success, but found %v", d.sql, err)
 		}
